netcf: hardcode --with-driver=redhat for target build
The configure script of netcf checks the HOST setup to identify the
network interface. This means it checks for Red Hat, SuSE, debian,
Ubuntu, etc. As this is an embedded cross build, it doesn't seem to make
sense to check the HOST in the first place to determine info for the
TARGET.

Signed-off-by: Ting Liu <b28495@freescale.com>
Change-Id: Ied75b9f553f92b8bbc6c3f050efff851fe2c2c63
Reviewed-on: http://git.am.freescale.net:8181/6901
Tested-by: Ting Liu <ting.liu@freescale.com>
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>